Untested Full Implementation
The code is now done, and completely untested. I need teh Orin cameras to be able to test it, and also to test it on the Orin. However, this code should be perfectly just fine and working. The Bash scripts are the only thing entirely untested and possibly not working correctly.